Python Examples doc added

+++ Second2.py
@@ -2,7 +2,7 @@ physics=80
 maths=100
 chemistry=100
 total=physics+maths+chemistry
-per=total/450*100
+per=int(total/450*100)
 print("Physics:",physics)
 print("Maths:",maths)
 print("Chemistry:",chemistry)
